+++ library/GitHub/Workflow/Command/Annotation.hs
@@ -48,6 +48,7 @@ module GitHub.Workflow.Command.Annotation
   , atColumn
 
     -- * Output
+  , MonadCommand (..)
   , ToCommand (..)
   , toCommand
   , ToByteString (..)
@@ -66,11 +67,7 @@ import GitHub.Workflow.Command.Annotation.Position.Columns
 import GitHub.Workflow.Command.Annotation.Position.Extent
 import GitHub.Workflow.Command.Annotation.Position.Line
 import GitHub.Workflow.Command.Annotation.Properties
-import GitHub.Workflow.Command.Syntax
-  ( FromMessage (..)
-  , Message (..)
-  , ToByteString (..)
-  , ToCommand (..)
-  , printByteStringLn
-  , toCommand
-  )
+import GitHub.Workflow.Command.Execution
+import GitHub.Workflow.Command.Syntax.Command
+import GitHub.Workflow.Command.Syntax.Message
+import GitHub.Workflow.Command.Syntax.ToByteString