+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-## [_Unreleased_](https://github.com/freckle/github-workflow-commands/compare/v0.0.0.0...main)
+## [_Unreleased_](https://github.com/freckle/github-workflow-commands/compare/v0.0.1.0...main)
+
+## [v0.0.1.0](https://github.com/freckle/github-workflow-commands/compare/v0.0.0.0...v0.0.1.0)
+
+Additions:
+
+- Re-exporting overview module `GitHub.Workflow.Command`, which is now the
+  primary module for users to import and find documentation
+- Class `MonadCommand` which is now the recommended way to execute commands
+- Support for `group`, `add-mask`, `stop-commands` commands
 
 ## [v0.0.0.0](https://github.com/freckle/github-workflow-commands/tree/v0.0.0.0)